requirements.txt
streamlit==1.31.0
openai==1.12.0
python-dotenv==1.0.1
Pillow==10.2.0
PyMuPDF==1.23.8
pdf2image==1.16.3